Template Pages
I have copied the index.tsx page to create a starting point for the front end of each page.

Matching page, messages page, temp login page, admin dashboard..

At the moment the login.tsx is a bit bugged so I created a nonfunctional example type page.

I will continue to mess around with the front end stuff but this is a starting point for now.